+++ src/extra/proposal/__init__.py
@@ -1,1 +1,4 @@
 from extra_proposal import *
+from extra_proposal.proposal import RunReference
+
+__all__ = ["Proposal", "RunReference"]